requirements.txt
# Core dependencies
crewai==0.28.8
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
python-multipart==0.0.6
jinja2==3.1.2

# AI and API dependencies
openai==1.6.1
elevenlabs==0.2.26
anthropic==0.7.8

# Video processing
moviepy==1.0.3
ffmpeg-python==0.2.0
Pillow==10.1.0

# Database and storage
supabase==2.0.2
boto3==1.34.0
sqlalchemy==2.0.23
alembic==1.13.1

# Background tasks
celery==5.3.4
redis==5.0.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0

# Additional dependencies
langchain==0.1.0
asyncio-mqtt==0.16.1
aiofiles==23.2.1 